+++ dash/main.py
@@ -1,8 +1,6 @@
 from fastapi.middleware.wsgi import WSGIMiddleware
 
-from api import app as api
+from api.api import app
 from dash_app import app as dashboard1
-
-app = api
 
 app.mount("", WSGIMiddleware(dashboard1.server))