+++ django_api/napapi/reference_manager/serializers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from .models import Reference
-
-
-class ReferenceSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Reference
-        fields = ["wid", "created", "data"]