+++ pages/visible_structure_search.py
@@ -5,7 +5,7 @@ import dash_bootstrap_components as dbc
 from dash import Input, Output, callback, dcc
 
 import plotly_dash_ketcher
-from config import PAGE_SIZE
+from dash_config import PAGE_SIZE
 from dash_common import generate_structures_cards
 from model import DataModel