+++ update/generate_database_taxo.py
@@ -2,7 +2,6 @@
 import csv
 import logging
 import pickle
-from io import StringIO
 from pathlib import Path
 
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
@@ -43,61 +42,65 @@ def run(path: Path) -> None:
 
         logging.info(f" Found {len(taxon_direct_parents)} taxa")
 
-    reader = csv.reader(StringIO(str(path / "taxa_parents.csv")))
-    reader.__next__()
-    for line in reader:
-        (
-            taxon_id,
-            taxon_name,
-            taxon_rank_id,
-            relative_id,
-            relative_name,
-            relative_rank,
-            distance,
-        ) = line
-        taxon_id = int(taxon_id)
-        relative_id = int(relative_id)
-        taxon_rank_id = int(taxon_rank_id)
-        distance = int(distance)
+    with open(path / "taxa_parents.csv", "r") as f:
+        reader = csv.reader(f)
+        next(reader)
+        # Todo probably replace with a dict csv reader
+        for line in reader:
+            (
+                taxon_id,
+                taxon_name,
+                taxon_rank_id,
+                relative_id,
+                relative_name,
+                relative_rank,
+                distance,
+            ) = line
+            taxon_id = int(taxon_id)
+            relative_id = int(relative_id)
+            taxon_rank_id = int(taxon_rank_id)
+            distance = int(distance)
 
-        if relative_id not in taxon_children:
-            taxon_children[relative_id] = set()
-        if taxon_id not in taxon_direct_parents:
-            taxon_direct_parents[taxon_id] = set()
-        if taxon_id not in taxon_names:
-            taxon_names[taxon_id] = taxon_name
-        if relative_id not in taxon_names:
-            taxon_names[relative_id] = relative_name
-        if taxon_id not in taxon_ranks:
-            taxon_ranks[taxon_id] = set()
-        if relative_id not in taxon_ranks:
-            taxon_ranks[relative_id] = set()
-        if taxon_id not in taxon_parents_with_distance:
-            taxon_parents_with_distance[taxon_id] = {}
+            if relative_id not in taxon_children:
+                taxon_children[relative_id] = set()
+            if taxon_id not in taxon_direct_parents:
+                taxon_direct_parents[taxon_id] = set()
+            if taxon_id not in taxon_names:
+                taxon_names[taxon_id] = taxon_name
+            if relative_id not in taxon_names:
+                taxon_names[relative_id] = relative_name
+            if taxon_id not in taxon_ranks:
+                taxon_ranks[taxon_id] = set()
+            if relative_id not in taxon_ranks:
+                taxon_ranks[relative_id] = set()
+            if taxon_id not in taxon_parents_with_distance:
+                taxon_parents_with_distance[taxon_id] = {}
 
-        if distance == 1:
-            taxon_direct_parents[taxon_id].add(relative_id)
+            if distance == 1:
+                taxon_direct_parents[taxon_id].add(relative_id)
 
-        taxon_children[relative_id].add(taxon_id)
-            # We also add the children of the ones above
-        if taxon_id in taxon_children:
-            for child_id in taxon_children[taxon_id]:
-                taxon_children[relative_id].add(child_id)
+            taxon_children[relative_id].add(taxon_id)
+                # We also add the children of the ones above
+            if taxon_id in taxon_children:
+                for child_id in taxon_children[taxon_id]:
+                    taxon_children[relative_id].add(child_id)
 
-        taxon_ranks[taxon_id].add(taxon_rank_id)
+            taxon_ranks[taxon_id].add(taxon_rank_id)
 
-        taxon_ranks[relative_id].add(relative_rank)
+            taxon_ranks[relative_id].add(relative_rank)
 
-        taxon_parents_with_distance[taxon_id][relative_id] = distance
+            taxon_parents_with_distance[taxon_id][relative_id] = distance
 
-    reader = csv.reader(StringIO(str(path / "taxa_ranks.csv")))
-    reader.__next__()
-    for line in reader:
-        ranks_names[int(line[0])] = line[1]
+    with open(path / "taxa_ranks.csv", "r") as f:
+        reader = csv.reader(f)
+        next(reader)
+        for line in reader:
+            ranks_names[int(line[0])] = line[1]
 
-    reader = csv.reader(StringIO(str(path / "taxa_all.csv")))
-    reader.__next__()
-    dict_all_taxa: dict = {i[0]: i[1] for i in reader}
+    with open(path / "taxa_all.csv", "r") as f:
+        reader = csv.reader(f)
+        next(reader)
+        dict_all_taxa: dict = {i[0]: i[1] for i in reader}
 
     database = {
         "taxonomy_direct_parents": taxon_direct_parents,